Header logo added and index.js bug fixed
